The changes in this commit are the first set of changes to make developing Ramaze itself a bit easier for everybody. Some of the changes may not fit the idea of others such as @manveru so feel free to discuss them.
First of all the Gemspec generation process has been changed. Rather than using
a Rake task to build/update the Gemspec Ramaze now has it's own manually created
Gemspec. The dependencies of this Gemspec are specified in Ramaze::DEPENDENCIES
and Ramaze::DEVELOPMENT_DEPENDENCIES, these files are also used for the setup
task.

Second I've removed several Rake tasks that appeared to be useless or did not
fit the idea of Ramaze (the jQuery task). I've also removed the Git tasks and
some other tasks that did not add anything useful to the development process.

If any of these changes do not meet any particular standards or ideas feel free
to discuss them but note that this is the first commit of a series so things
might change a bit after this commit has been pushed.