Second commit of the series. In this commit I removed the benchmark/ directory and made some final changes to the building/releasing process of Ramaze.
The benchmark/ directory was removed as I personally feel benchmarks should be
done by a third-party that has an unbiased opinion of Ramaze, on top of that I
doubt if most code in the benchmark still applies to today's Ramaze.

The building/installing process has also been finished. Building a gemspec is no
longer possible (as it's not needed) but building an actual gem file and
releasing it is still possible. Building a gem file can be done as following:

    $ rake gem:build
    $ rake gem:install

Releasing a version of Ramaze on Rubygems can be done as following:

    $ rake release