Gave the Ramaze executable the much needed love.
A few things have changed to the executable so far. The biggest change is that
rather than using regular expressions to "parse" all given options it now uses
OptionParser. This comes with the added advantage of making it easy to set
banners, options, running commands and so on. Another big change is that rather
than dumping most of the code in a single file each command is located in it's
own file making it easier to maintain them.

Most of the code in this commit is still backwards compatible and should still
work on Windows based system (since I just used the code written by @bougyman for
that). There are however a few minor changes. Because OptionParser is used the
output displayed when specifying -h or --help has changed, this however is a
good thing as the output now follows the Unix standards of command line
executables.

Another thing that has changed in terms of command behavior is that most
commands related to managing applications either accept a directory or a file.
For example, the command $ ramaze start now allows you to specify a directory
(or file) that points to the location of a Rackup configuration file. This is
actually quite useful as it makes it possible to run the ramaze executable from
any given location.

In order to make it easier to spec the executable and it's commands (I have yet
to begin working on this) I decided to manually specify all the possible Rack
and Ruby options the command Ramaze::Bin::Start accepts. Before this commit
these were extracted by parsing the output of rackup -h, this however proved to
be quite annoying as with every change of this output our specs would break
(even if it would be something simple such as newlines being removed).

A few examples of the new executable setup:

    $ ramaze start
    $ ramaze start /home/someuser/ramaze/
    $ ramaze start /home/someuser/ramaze/config.ru

Debugging using IRB:

    $ ramaze console
    $ ramaze console /home/someuser/ramaze
    $ ramaze console /home/someuser/ramaze/start.rb

Or perhaps viewing the status of an application:

    $ ramaze status
    $ ramaze status /home/someuser/ramaze -P /home/someuser/ramaze/ramaze.pid

For more information see the files located in lib/ramaze/bin, the old files
located in lib/ramaze/tool will be removed in the next commit.