Modified the layout helper so that it no longer overwrites layout().
Do note that this change has a minor limitation: it will only create a "backup"
of the existing layout (located in a trait called :layout) the first time
set_layout() is called. After this it will just use the one that was set in the
first call.

This means that you can do the following and it would just work fine:

    layout     :default
    set_layout :alternative => [:method_1]

This would result in the layout "default" being use by default and the layout
"alternative" being used for method_1(). Before this commit this would result
in all methods except method_1() being rendered *without* a layout at all.

In the following example both calls to set_layout() will use the layout set by
layout() as a backup, the second one will not overwrite it but instead will also
use the backup:

    layout     :default
    set_layout :alternative => [:method_1]
    set_layout :another_one => [:method_2]

To cut a long story short, existing layouts are now properly used as a fallback
in case set_layout() doesn't result in any matching layouts for a particular
method.