Docs/style changes for the Redis cache.
The Redis cache has been documented and I've made some slight modifications to
it so that it works similar to the Sequel/Memcache cache in terms of setting
options. I've also modified the Sequel, Memcache and Redis cache classes to
use Ramaze.setup() for their dependencies, this should make it a bit easier for
users to use these caches.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>