Jruby support and simplified various tests.
In order to achieve jruby compatibility I had to make some modifications to the
development dependencies setup we've been using for a while now. I added a list
of unsupported gems (such as lokar) which will be "marked" as supported on
certain systems. For example, lokar is only supported if the user isn't running
Rubinius.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>