Re-generate CSRF tokens for valid requests.
Re-generating the CSRF tokens on each valid request fixes the annoying issue of
the tokens *always* expiring after 15 minutes. This is very annoying if you're
trying to edit some content and all of a sudden you're unable to submit a form
as the token has expired.

See #27 for more information.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>