Fixes blue_form html doc tags and code syntax
HTML tags referenced in the documentation were not escaped. Thus, when
rendered as HTML, those tags weren't displayed properly in the browser.
This patch escapes them as suggested by yorickpeterse.

One of the examples wasn't syntactically correct, and had an extra do
before the block. This is not removed and the example parses properly.