Adds blue form examples and #g hint
It wasn't obvious how one had to do to generate tags not covered by
current Form methods. This patch gives a hint on using Form#g to get a
Gestalt instance back, and use it as usual.

An example is added to illustrate this and show how to use additional
arguments to set attibutes.