Make Ramaze::Cache app aware.
Innate used a hardcoded value of "pristine" for cache namespaces due to not
having a concept of applications such as Ramaze has. This meant that if you had
two applications running on the same host and user as well as using the same
cache (e.g. Memcached) you'd end up with the problem of data being overwritten
by these applications.

For example, assume you have two applications:

1. "a"
2. "b"

If you were to run these applications on host "myhost" and user "myuser" you'd
end up with the following cache keys, regardless of the value of
`Ramaze.options.app.name`:

1. myhost-myuser-pristine-memcached
2. myhost-myuser-pristine-memcached

This commit will change those keys to the following:

1. myhost-myuser-a-memcached
2. myhost-myuser-b-memcached

Of course this assumes you've properly set the application names. If these are
not set the code will default to "pristine".

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>