Cleaning up shop.
To make it easier for developers to hack on Ramaze I've started changing
several parts of Ramaze related to how it handles dependencies and how the
release process is executed.

To make things easier I've added a Gemfile to the project (as well as an
updated .gems file). This means people can install the development gems by
simply running `bundle install`.

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>