New setup for writing Ramaze applications.
Most of the code from the `ramaze` command has been removed from the executable
into individual Rake tasks located in the prototype. This makes it easier for
developers to customize these parts (e.g. by using a custom REPL) as well as
making it easier to maintain Ramaze itself.

With these changes I've also removed the tutorials section from the userguide.
It hasn't been updated in a while and with these changes was no longer really
useful.

In short the following changes:

* ramaze start   => rake ramaze:start
* ramaze console => rake ramaze:irb or rake ramaze:pry
* Rake is now a dependency of Ramaze so that it's installed when installing
  Ramaze
* ramaze stop, status and restart have been removed

Signed-off-by: Yorick Peterse <yorickpeterse@gmail.com>