Adds more documentation about flash and flashbox
Added documentation on how flash can be used in view or layouts,
explaining that any key can be used.
An example shows how to take advantage of that to use flash keys as css
selectors.
Finally, a small example of how flashbox is used in views is shown.
Hopefully, this fixes #64